수정

+++ legacy/2021NIA/NiaIpSdnToAiDataLinkageApplication/src/main/java/com/nia/data/linkage/ipsdn/mapper/ipsdn/IpsdnDataMapper.java
@@ -20,7 +20,7 @@ public interface IpsdnDataMapper {
     ArrayList<NodeFactorVo> selectNodeFactorLink(int dataKey);
     ArrayList<LinkTrafficVo> selectTrafficList(int dataKey);
     ArrayList<SyslogDataVo> selectSyslogData(int dataKey);
-    ArrayList<SflowDataVo> selectSflowData(int dataKey);
+    ArrayList<SflowDataVo> selectSflowData(long dataKey);
 
     ArrayList<AlarmDataVo> selectAlarmList(int alarmno);
